refactor: Replace `NoteroItem` class with simple functions
